refactor: move block uniform handler into kingly
