+++ packages/charts/src/chart_types/flame_chart/render/ensure_webgl.ts
@@ -8,6 +8,7 @@
 
 import {
   bindVertexArray,
+  blockUniforms,
   createCompiledShader,
   createLinkedProgram,
   getAttributes,
@@ -15,34 +16,8 @@ import {
   resetState,
 } from '../../../common/kingly';
 import { GL } from '../../../common/webgl_constants';
-import { attributeLocations, colorFrag, roundedRectVert, roundedRectFrag, simpleRectVert } from '../shaders';
+import { attributeLocations, colorFrag, roundedRectFrag, roundedRectVert, simpleRectVert } from '../shaders';
 import { GLResources, NULL_GL_RESOURCES } from '../types';
-
-function blockUniforms(
-  gl: WebGL2RenderingContext,
-  uboVariableNames: string[],
-  [program, ...otherPrograms]: WebGLProgram[],
-) {
-  const blockIndex = gl.getUniformBlockIndex(program, 'Settings');
-  const blockSize = gl.getActiveUniformBlockParameter(program, blockIndex, GL.UNIFORM_BLOCK_DATA_SIZE);
-  const uboBuffer = gl.createBuffer();
-  if (uboBuffer === null) throw new Error('Whoa, could not create uboBuffer');
-  gl.bindBuffer(gl.UNIFORM_BUFFER, uboBuffer);
-  gl.bufferData(gl.UNIFORM_BUFFER, blockSize, gl.DYNAMIC_DRAW);
-  // gl.bindBuffer(gl.UNIFORM_BUFFER, null);
-  gl.bindBufferBase(gl.UNIFORM_BUFFER, 0, uboBuffer);
-  const uboVariableIndices = gl.getUniformIndices(program, uboVariableNames);
-  if (uboVariableIndices === null) throw new Error('Whoa, could not get uboVariableIndices');
-  const uboVariableOffsets = gl.getActiveUniforms(program, uboVariableIndices, gl.UNIFORM_OFFSET);
-  const uniforms = new Map(
-    uboVariableNames.map((name, i) => [name, { index: uboVariableIndices[i], offset: uboVariableOffsets[i] }]),
-  );
-
-  // per program part
-  [program, ...otherPrograms].forEach((p) => gl.uniformBlockBinding(p, gl.getUniformBlockIndex(p, 'Settings'), 0));
-
-  return { uboBuffer, uniforms };
-}
 
 /** @internal */
 export function ensureWebgl(
@@ -81,24 +56,23 @@ export function ensureWebgl(
     attributeLocations,
   );
 
-  /**
-   * Uniform blocks
-   */
-
-  const uboVariableNames = [
-    'focus',
-    'resolution',
-    'gapPx',
-    'minFillRatio',
-    'rowHeight0',
-    'rowHeight1',
-    't',
-    'cornerRadiusPx',
-    'hoverIndex',
-    'pickLayer',
-  ];
-
-  const { uboBuffer, uniforms } = blockUniforms(gl, uboVariableNames, [geomProgram, pickProgram]);
+  const { uboBuffer, uniforms } = blockUniforms(
+    gl,
+    'Settings',
+    [
+      'focus',
+      'resolution',
+      'gapPx',
+      'minFillRatio',
+      'rowHeight0',
+      'rowHeight1',
+      't',
+      'cornerRadiusPx',
+      'hoverIndex',
+      'pickLayer',
+    ],
+    [geomProgram, pickProgram],
+  );
 
   /**
    * Resource allocation: Render setup
